+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "24.0.1"
 
     defaultConfig {
-        applicationId "io.soramitsu.irohaapiwrapper"
+        applicationId "io.soramitsu.iroha"
         minSdkVersion 19
         targetSdkVersion 24
         versionCode 1
